Refactor UserDao implementation
